Move unit tests to dedicated files
